raspi/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
paho-mqtt==2.1.0
rpi-lgpio==0.6
pillow==11.0.0